2.5.7
- Fixed SyncList deserializing improperly.
- Fixed NullReferenceException in PredictedObject.
- Added PredictedObject.ResendType.
- Fixed PredictedObject spectator smoothing sometimes fighting against owner smoothing.

+++ Assets/FishNet/Example/All/CustomSyncType/Custom Struct Sync/StructySync.cs
@@ -148,7 +148,7 @@ namespace FishNet.Example.CustomSyncObject
         public override void WriteDelta(PooledWriter writer, bool resetSyncTick = true)
         {
             base.WriteDelta(writer, resetSyncTick);
-            writer.WriteUInt32((uint)_changed.Count);
+            writer.WriteInt32(_changed.Count);
 
             for (int i = 0; i < _changed.Count; i++)
             {
@@ -201,7 +201,7 @@ namespace FishNet.Example.CustomSyncObject
             * and potentially overwrite data not yet sent. */
             bool asClientAndHost = (!asServer && base.NetworkManager.IsServer);
 
-            int changes = (int)reader.ReadUInt32();
+            int changes = reader.ReadInt32();
             for (int i = 0; i < changes; i++)
             {
                 CustomOperation operation = (CustomOperation)reader.ReadByte();
