+++ Assets/FishNet/Runtime/Object/Synchronizing/SyncDictionary.cs
@@ -245,7 +245,7 @@ namespace FishNet.Object.Synchronizing
             {
                 //False for not full write.
                 writer.WriteBoolean(false);
-                writer.WriteUInt32((uint)_changed.Count);
+                writer.WriteInt32(_changed.Count);
 
                 for (int i = 0; i < _changed.Count; i++)
                 {
@@ -285,7 +285,7 @@ namespace FishNet.Object.Synchronizing
             base.WriteHeader(writer, false);
             //True for full write.
             writer.WriteBoolean(true);
-            writer.WriteUInt32((uint)Collection.Count);
+            writer.WriteInt32(Collection.Count);
             foreach (KeyValuePair<TKey, TValue> item in Collection)
             {
                 writer.WriteByte((byte)SyncDictionaryOperation.Add);
@@ -318,7 +318,7 @@ namespace FishNet.Object.Synchronizing
             if (fullWrite)
                 collection.Clear();
 
-            int changes = (int)reader.ReadUInt32();
+            int changes = reader.ReadInt32();
             for (int i = 0; i < changes; i++)
             {
                 SyncDictionaryOperation operation = (SyncDictionaryOperation)reader.ReadByte();
